Moved create_test_rtc_nc_product() to metadata_utils.py so it can be imported from within a built container
